Enhanced the handling of ID and serialization of structs.
Added special RepositoryDefiniton property (bool flag) "IsCustomID"
to indicate if the handling of the "id" properties is left to the
library (```default``` behavior like mapping "id" => "_id" for MongoDB)
or to the developer (```custom``` behaviour where the developer has the
responsibility to manage the object "id").

Fixed the serialization of Structs for annotated (tag) fields - the proper
names are looked up in the filed tag - first "bson" then "json" if present.