Changed the returned errors type in helper.go to ErrInvalidInput.
All of the errors that are changed are caused by providing invalid input
to the helper function. This should cause an ```ErrInvalidInput``` error
to be returned, instead of generic error. This way we can test easily and
respond back to the user with proper error type and code (for example
HTTP 400 Bad Request).